requirements.txt
fpdf2
requests
qrcode